+++ changelog.md
@@ -1,6 +1,42 @@
 # Changelog
 All notable changes to this project will be documented in this file. Major releases are documented [here](https://github.com/jgyates/genmon/releases)
 
+## V1.13.29 - 2019-08-20
+- Fixed a few typos
+
+## V1.13.28 - 2019-07-20
+- Minor update to gentankutil.py
+- Added add-on for Amazon Alexa
+
+## V1.13.27 - 2019-07-20
+- Updated icon on add-on page
+- Added fuel metric to Maintenance page for models that support fuel monitoring
+
+## V1.13.26 - 2019-07-19
+- Added email notification for Evolution 2 firmware
+
+## V1.13.25 - 2019-07-12
+- Added initial support for fuel consumption for Evolution Air Cooled Natural Gas units. Note: NG 30 day fuel consumption is located on the maintenance page. Fuel gauge is supported for NG.
+
+## V1.13.24 - 2019-07-10
+- Added URL to logout if using secure web settings: https://addressofpi/logout
+
+## V1.13.23 - 2019-07-07
+- Changes to all unicode comments to be stored in service journal
+
+## V1.13.22 - 2019-07-06
+- Added config file options to set nominal line voltage values for non US grids (Evo and Nexus only)
+
+## V1.13.21 - 2019-06-28
+- Slight modification to the logic used when detecting the type of alarm for Evolution controllers due to the last alarm code register not updating after a Service Due alarm.
+
+## V1.13.20 - 2019-06-26
+- Added option "Ignore Unknown Values" that will ignore unknown values of modbus register 0001 for Evolution 2.0 controllers. This setting is in the advanced section in the web interface.
+
+## V1.13.19 - 2019-06-25
+- Added more logging for better diagnostics when fatal errors occur (e.g. invalid serial port specified)
+- Corrected one problem with invalid file name for log file when sending logs
+
 ## V1.13.18 - 2019-06-19
 - Fixed problem in gentankutil.py